# Conflicts:
#	.github/workflows/abacus-playwright-tests.yml
#	.github/workflows/vulnerability.yml
#	.github/workflows/xtest.yml
#	opentdf.Tiltfile

+++ abacus-integration-test/e2e/Authorities.spec.ts
@@ -6,7 +6,7 @@ import {
     assertAttributeCreatedMsg,
     getAccessToken,
     deleteAuthorityViaAPI,
-    deleteAttributeViaAPI
+    removeAllAttributesOfAuthority,
 } from './helpers/operations';
 import { test } from './helpers/fixtures';
 import { selectors } from "./helpers/selectors";
@@ -17,11 +17,11 @@ let apiContext: APIRequestContext;
 test.describe('<Authorities/>', () => {
     test.beforeEach(async ({ page , playwright, authority}) => {
         await authorize(page);
-        authToken = await getAccessToken(page)
+        authToken = await getAccessToken(page);
 
-        await page.goto('/attributes');
-        // click the token message to close it and overcome potential overlapping problem
-        await page.locator(selectors.tokenMessage).click()
+        await page.getByRole('link', { name: 'Attributes' }).click();
+        await page.waitForURL('**/attributes');
+
         await createAuthority(page, authority);
         // click success message to close it and overcome potential overlapping problem
         const authorityCreatedMsg = page.locator(selectors.alertMessage, {hasText:'Authority was created'})
@@ -34,37 +34,47 @@ test.describe('<Authorities/>', () => {
         });
     });
 
+    test.afterEach(async ({ authority}, testInfo) => {
+        // Because authority in this test already deleted
+        await removeAllAttributesOfAuthority(apiContext, authority);
+        if (testInfo.title !== 'delete authority if there are no assigned attributes') {
+            await deleteAuthorityViaAPI(apiContext, authority);
+        }
+    })
+
     test.afterAll(async ({ }) => {
         await apiContext.dispose();
     });
 
     test('renders initially', async ({ page, authority}) => {
-        await page.goto('/authorities');
+        await page.getByRole('link', { name: 'Authorities' }).click();
+        await page.waitForURL('**/authorities');
+
         const header = page.locator(selectors.authoritiesPage.header, { hasText: "Authorities" });
         await expect(header).toBeVisible();
-
-        await test.step('Cleanup', async() => {
-            await deleteAuthorityViaAPI(apiContext, authority)
-        })
     });
 
     test('delete authority if there are no assigned attributes', async ({ page, authority}) => {
-        await page.goto('/authorities');
-        // click the token message to close it and overcome potential overlapping problem
-        await page.locator(selectors.tokenMessage).click()
+        await test.step('Open authorities route', async () => {
+            await page.getByRole('link', { name: 'Authorities' }).click();
+            await page.waitForURL('**/authorities');
+        });
 
-        const originalTableRows = await page.$$(selectors.authoritiesPage.authoritiesTableRow)
+        await page.waitForSelector(selectors.authoritiesPage.authoritiesTableRow);
+        const originalTableRows = await page.locator(selectors.authoritiesPage.authoritiesTableRow).all();
         const originalTableSize = originalTableRows.length
 
-        const deleteAuthorityButtonForTheLastRowItem = await page.locator('#delete-authority-button >> nth=-1')
-        await deleteAuthorityButtonForTheLastRowItem.click()
+        const deleteButton = await page.getByRole('row', { name: `${authority} Delete` }).getByRole('button', { name: 'Delete' });
+        await deleteButton.click();
 
-        await test.step('Should be able to close the dialog and cancel authority removal', async() => {
+        await test.step('Should be able to close the dialog and cancel authority removal', async () => {
             await page.click(selectors.authoritiesPage.confirmDeletionModal.cancelDeletionBtn)
         })
 
-        await deleteAuthorityButtonForTheLastRowItem.click()
-        await page.click(selectors.authoritiesPage.confirmDeletionModal.confirmDeletionBtn)
+        await test.step('Confirm Deletion Modal', async () => {
+            await deleteButton.click();
+            await page.click(selectors.authoritiesPage.confirmDeletionModal.confirmDeletionBtn)
+        });
 
         await test.step('Assert success message', async() => {
             const successfulDeletionMsg = await page.locator(selectors.alertMessage, {hasText: `Authority ${authority} deleted`})
@@ -72,7 +82,7 @@ test.describe('<Authorities/>', () => {
             await successfulDeletionMsg.click()
         })
 
-        const updatedTableRows = await page.$$(selectors.authoritiesPage.authoritiesTableRow)
+        const updatedTableRows = await page.locator(selectors.authoritiesPage.authoritiesTableRow).all();
         const updatedTableSize = updatedTableRows.length
 
         expect(updatedTableSize === (originalTableSize - 1)).toBeTruthy()
@@ -81,9 +91,9 @@ test.describe('<Authorities/>', () => {
     test('Authority removal is failed when contains assigned attributes', async ({ page, authority, attributeName, attributeValue}) => {
         await createAttribute(page, attributeName, [attributeValue])
         await assertAttributeCreatedMsg(page)
-        await page.goto('/authorities');
-        // click the token message to close it and overcome potential overlapping problem
-        await page.locator(selectors.tokenMessage).click()
+
+        await page.getByRole('link', { name: 'Authorities' }).click();
+        await page.waitForURL('**/authorities');
 
         const deleteAuthorityButtonForTheLastRowItem = await page.locator('#delete-authority-button >> nth=-1')
         await deleteAuthorityButtonForTheLastRowItem.click()
@@ -93,11 +103,6 @@ test.describe('<Authorities/>', () => {
             const removalFailedMsg = await page.locator(selectors.alertMessage, {hasText: `Something went wrong`})
             await expect(removalFailedMsg).toBeVisible()
             await removalFailedMsg.click()
-        })
-
-        await test.step('Cleanup', async() => {
-            await deleteAttributeViaAPI(apiContext, authority, attributeName, [attributeValue])
-            await deleteAuthorityViaAPI(apiContext, authority)
         })
     });
 });
